Fix OD params
